OSINT/WhoIs/requirements.txt
python-whois
dnspython
